refactor: fix MSVC warnings

+++ include/soundswallower/hash_table.h
@@ -216,12 +216,9 @@ void *hash_table_enter(hash_table_t *h, /**< In: Handle of hash table in which t
 
 /**
  * Add a 32-bit integer value to a hash table.
- *
- * This macro is the clean way to do this and avoid compiler warnings
- * on 64-bit platforms.
  */
 #define hash_table_enter_int32(h,k,v) \
-    ((int32)(long)hash_table_enter((h),(k),(void *)(long)(v)))
+    ((int32)hash_table_enter((h),(k),(void *)(v)))
 
 /**
  * Add a new entry with given key and value to hash table h.  If the
@@ -244,12 +241,9 @@ void *hash_table_replace(hash_table_t *h, /**< In: Handle of hash table in which
 
 /**
  * Replace a 32-bit integer value in a hash table.
- *
- * This macro is the clean way to do this and avoid compiler warnings
- * on 64-bit platforms.
  */
 #define hash_table_replace_int32(h,k,v) \
-    ((int32)(long)hash_table_replace((h),(k),(void *)(long)(v)))
+    ((int32)hash_table_replace((h),(k),(void *)(v)))
 
 /**
  * Delete an entry with given key and associated value to hash table
